# Change format of MixedAmount so it is normalised by design