This PR changes the internal format of MixedAmount to a Strict Map, which is normalised by construction. This format is used internally in `normaliseHelper` already. Anything which should not be normalised should be stored as an `[Amount]`.

I'll add performance numbers here, but I'll wait until #1505 is merged to add those to the commit messages, since the improvements will interfere with one another.