@Xitian9 2021-03-22
Here are the preliminary performance numbers (pre-#1505). All other metrics and tests that I've checked produce changes that are small enough that we can reasonably assume no effect.

`hledger reg -f examples/10000x10000x10.journal`
- Total memory in use drops from 189MiB to 167MiB, a 12% drop

`hledger bal -f examples/10000x10000x10.journal`
- Total bytes allocated in the heap drops from 26.9GB to 13.1GB, a 51% drop
- Total (profiled) time drops from 21.8s to 16.0s, a 27% drop

`hledger bal -f examples/10000x10000x10.journal -M`
- Total bytes allocated in the heap drops from 195.5GB to 53.9GB, a 72% drop
- Total (profiled) time drops from 93.4s to 48.6s, a 48% drop