@Xitian9 2021-03-22
Running the above tests with the more realistically sized `examples/10000x1000x10.journal` produced smaller but consistent results (except I'm getting a slightly higher “total memory in use” stat for `hledger bal -M`, which is not reproduced for `hledger bal`).