@Xitian9 2021-03-22
This also adds some tests for JSON output, at least partially addressing #1360.