@Xitian9 2021-03-24
I've added performance comments in the commit which changes the internal representation of MixedAmount (at time of writing, the second commit in this PR). Overall there's a large improvement for journals with lots of accounts, and a smaller but still significant improvement for journals with a more reasonable number of accounts.