@simonmichael 2021-04-06
Some casual time benchmarking: 

Simple reports on a 15 year journal - a slight improvement:
```
$ hledger.1509 -f ~/adm/all.journal stats | grep -iE '(accounts|transactions) +:'
Transactions             : 15119 (2.9 per day)
Accounts                 : 1678 (depth 8)
$ quickbench "hledger.pre1509 -f all.journal stats" "hledger.1509 -f all.journal stats"
Running 2 tests 1 times at 2021-04-05 16:07:53 HST:

Best times:
+--------------------------------------++------+
|                                      ||      |
+======================================++======+
| hledger.pre1509 -f all.journal stats || 2.39 |
| hledger.1509 -f all.journal stats    || 2.21 |
+--------------------------------------++------+
$ quickbench "hledger.pre1509 -f all.journal bal" "hledger.1509 -f all.journal bal"
Running 2 tests 1 times at 2021-04-05 16:08:04 HST:

Best times:
+------------------------------------++------+
|                                    ||      |
+====================================++======+
| hledger.pre1509 -f all.journal bal || 2.41 |
| hledger.1509 -f all.journal bal    || 2.16 |
+------------------------------------++------+
```

Two runs of the standard benchmarks - ~10% faster print, ~15% slower register, ~10-25% faster balance (but not with large files perhaps ?):
```
$ BENCHEXES=hledger.pre1509,hledger.1509 make bench

run quick performance benchmarks in bench.sh (approximate, can be skewed):
Running 5 tests 1 times with 2 executables at 2021-04-05 16:08:27 HST:

Best times:
+----------------------------------------------------++-----------------+--------------+
|                                                    || hledger.pre1509 | hledger.1509 |
+====================================================++=================+==============+
| -f examples/10000x1000x10.journal print            ||            0.97 |         0.86 |
| -f examples/10000x1000x10.journal register         ||            8.02 |         8.61 |
| -f examples/10000x1000x10.journal balance          ||            1.24 |         0.86 |
| -f examples/1000x1000x10.journal balance --weekly  ||            0.83 |         0.68 |
| -f examples/10000x1000x10.journal balance --weekly ||           11.04 |        11.40 |
+----------------------------------------------------++-----------------+--------------+
$ BENCHEXES=hledger.pre1509,hledger.1509 make bench

run quick performance benchmarks in bench.sh (approximate, can be skewed):
Running 5 tests 1 times with 2 executables at 2021-04-05 16:20:59 HST:

Best times:
+----------------------------------------------------++-----------------+--------------+
|                                                    || hledger.pre1509 | hledger.1509 |
+====================================================++=================+==============+
| -f examples/10000x1000x10.journal print            ||            1.02 |         0.87 |
| -f examples/10000x1000x10.journal register         ||            7.56 |         8.57 |
| -f examples/10000x1000x10.journal balance          ||            1.20 |         0.90 |
| -f examples/1000x1000x10.journal balance --weekly  ||            0.86 |         0.72 |
| -f examples/10000x1000x10.journal balance --weekly ||           11.01 |        10.51 |
+----------------------------------------------------++-----------------+--------------+
```

Two runs of your benchmarks - register is a little faster, balance is >100% faster (2x), balance -M ~30-40% faster:
```
$ quickbench -w hledger.pre1509,hledger.1509 "hledger reg -f examples/10000x10000x10.journal" "hledger bal -f examples/10000x10000x10.journal" "hledger bal -f examples/10000x10000x10.journal -M"
Running 3 tests 1 times with 2 executables at 2021-04-05 16:14:56 HST:

Best times:
+-------------------------------------------++-----------------+--------------+
|                                           || hledger.pre1509 | hledger.1509 |
+===========================================++=================+==============+
| reg -f examples/10000x10000x10.journal    ||            8.03 |         7.98 |
| bal -f examples/10000x10000x10.journal    ||            3.03 |         1.29 |
| bal -f examples/10000x10000x10.journal -M ||           36.63 |        24.79 |
+-------------------------------------------++-----------------+--------------+
$ quickbench -w hledger.pre1509,hledger.1509 "hledger reg -f examples/10000x10000x10.journal" "hledger bal -f examples/10000x10000x10.journal" "hledger bal -f examples/10000x10000x10.journal -M"
Running 3 tests 1 times with 2 executables at 2021-04-05 16:17:03 HST:

Best times:
+-------------------------------------------++-----------------+--------------+
|                                           || hledger.pre1509 | hledger.1509 |
+===========================================++=================+==============+
| reg -f examples/10000x10000x10.journal    ||            9.48 |         8.47 |
| bal -f examples/10000x10000x10.journal    ||            5.22 |         2.07 |
| bal -f examples/10000x10000x10.journal -M ||           44.83 |        24.23 |
+-------------------------------------------++-----------------+--------------+
```