@Xitian9 2021-04-12
In the end I was able to use `MixedAmount` for the `pamount` field of Postings without any problem: whatever blocker there was originally must have gotten resolved. This also improves the performance numbers, which I've updated in the commit message. There is no longer anything that runs slower/uses more memory: it's all-around improvements. Those improvements are actually slightly better than before, but since we're also comparing against a better baseline the percentage improvements are slightly lower.

Let me know if you have any other questions.