@Xitian9 2021-04-12
And of course, the requisite build failure due to Semigroups in ghc8.2…